Select Potential Voxels

Takes the difference in T2 images, smoothes this difference, and then finds voxels greater than one SD of the smooothed mask as potential voxels and returns it

voxel_select(normalized_baseline_t2, normalized_follow_up_t2, brain_mask,
  sigma = diag(3, 3), ksize = 5)

Arguments

normalized_baseline_t2
Baseline T2 image, array or object class nifti that

normalized_follow_up_t2
Followup T2 image, array or object class nifti that

brain_mask
A 3D 0-1 mask that delimits where the smoothing occurs, passed to GaussSmoothArray

sigma
Sigma passed to GaussSmoothArray

ksize
Kernel size passed to GaussSmoothArray

Value

Array or object class nifti depending on imput iamges

See also